fix: removes debugger statement
refactor: changes mutation

refactor: changes url and variables

refactor: move postData to own utils dir
